rewrite fetchFromAPI fn
